Fix unit test cases for FreeRTOS_Sockets (#651)
* Fix unit test cases for FreeRTOS_Sockets_TCP_API_utest
* Fix unit test case for FreeRTOS_Sockets_privates_utest
* Fix unit test case for FreeRTOS_Sockets_GenericAPI_utest
* Fix unit test case for FreeRTOS_Sockets_UDP_API_utest
* Update ucASCIIToHex
