+++ source/FreeRTOS_IPv6_Sockets.c
@@ -526,50 +526,6 @@ static BaseType_t prv_inet_pton6_add_nibble( struct sPTON6_Set * pxSet,
 
 /**
  * @brief Convert an ASCII character to its corresponding hexadecimal value.
- *        Accepted characters are 0-9, a-f, and A-F.
- *
- * @param[in] cChar: The character to be converted.
- *
- * @return The hexadecimal value, between 0 and 15.
- *         When the character is not valid, socketINVALID_HEX_CHAR will be returned.
- */
-uint8_t ucASCIIToHex( char cChar )
-{
-    char cValue = cChar;
-    uint8_t ucNew;
-
-    if( ( cValue >= '0' ) && ( cValue <= '9' ) )
-    {
-        cValue -= ( char ) '0';
-        /* The value will be between 0 and 9. */
-        ucNew = ( uint8_t ) cValue;
-    }
-    else if( ( cValue >= 'a' ) && ( cValue <= 'f' ) )
-    {
-        cValue -= ( char ) 'a';
-        ucNew = ( uint8_t ) cValue;
-        /* The value will be between 10 and 15. */
-        ucNew += ( uint8_t ) 10;
-    }
-    else if( ( cValue >= 'A' ) && ( cValue <= 'F' ) )
-    {
-        cValue -= ( char ) 'A';
-        ucNew = ( uint8_t ) cValue;
-        /* The value will be between 10 and 15. */
-        ucNew += ( uint8_t ) 10;
-    }
-    else
-    {
-        /* The character does not represent a valid hex number, return 255. */
-        ucNew = ( uint8_t ) socketINVALID_HEX_CHAR;
-    }
-
-    return ucNew;
-}
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Convert an ASCII character to its corresponding hexadecimal value.
  *        A :: block was found, now fill in the zero's.
  * @param[in] pxSet : A set of variables describing the conversion.
  */
