+++ source/include/FreeRTOS_Sockets.h
@@ -480,6 +480,12 @@
                                      char * pcDestination,
                                      socklen_t uxSize );
 
+/**
+ * @brief Convert an ASCII character to its corresponding hexadecimal value.
+ *        Accepted characters are 0-9, a-f, and A-F.
+ */
+    uint8_t ucASCIIToHex( char cChar );
+
 /** @brief This function converts a human readable string, representing an 48-bit MAC address,
  * into a 6-byte address. Valid inputs are e.g. "62:48:5:83:A0:b2" and "0-12-34-fe-dc-ba". */
     BaseType_t FreeRTOS_EUI48_pton( const char * pcSource,
